requirements.txt
weaviate-client
transformers
torch
Pillow
sentence-transformers
langdetect
flask
fastapi
uvicorn[standard]
python-multipart
psycopg2-binary
celery
redis